% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseStrands.R
\name{collapseStrands}
\alias{collapseStrands}
\title{Collapse ratioFile and generate bedGraph}
\usage{
collapseStrands(ratioFile, fileName, verbose = TRUE, asBedgraph = FALSE)
}
\arguments{
\item{ratioFile}{A dataframe with wcRatio coordinates.}

\item{fileName}{Identifier used in bedgraph header}

\item{verbose}{Verbose messages if\code{TRUE}}

\item{asBedgraph}{if\code{TRUE} Formates output file as bedgraph}
}
\description{
Collapse space between values in ratioFile that are consecutive, with bedGraph format option
}
\author{
Ashley D. Sanders, Mark Hills
}

