% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFreqs.R
\name{countFreqs}
\alias{countFreqs}
\title{Calculate wcRatios for strandTable}
\usage{
countFreqs(strandTable, checkNum = 10, verbose = TRUE)
}
\arguments{
\item{strandTable}{a dataframe object of sequencing reads}

\item{checkNum}{Window size used to calculate wcRatios}

\item{verbose}{Verbose messages if \code{TRUE}.}
}
\description{
Calculate wcRatios for input bedfile using a read-based sliding window of checkNum size
}
\author{
Ashley D. Sanders, Mark Hills
}

