% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findROILocation.R
\name{findROILocation}
\alias{findROILocation}
\title{finds ROI locations in collapsed wcRatio file}
\usage{
findROILocation(collapseFile, fileFrequencies, chrState = 0, verbose = TRUE,
  baselineThreshold = 0.8, elongationPenalty = 3, minReads = 4)
}
\arguments{
\item{collapseFile}{Dataframe of collapsed wcRatios}

\item{fileFrequencies}{Bedfile of reads used to calculate wcRatios}

\item{chrState}{Identify ROIs that deviate from chrState}

\item{verbose}{Verbose messages if\code{TRUE}}

\item{baselineThreshold}{Minimum threshold used to call ROIs from wcRatio file}

\item{elongationPenalty}{Extends roiRowList}

\item{minReads}{Minimimum reads required to call the ROI}
}
\description{
locates putative inversions based on distribution of wcRatios and generates an ROI file
}
\author{
Ashley D. Sanders, Mark Hills
}

