% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgaps.R
\name{plotGaps}
\alias{plotGaps}
\title{Plots gaps in output png}
\usage{
plotGaps(gapfile, chr, startLoc, endLoc, padding)
}
\arguments{
\item{gapfile}{File with gap locations}

\item{chr}{Genomic coordinate chromosome}

\item{startLoc}{Genomic coordinate start position}

\item{endLoc}{Genomic coordinate end position}

\item{padding}{Number of bases to extend beyond genomic coordinates listed in regionTable}

\item{binSize}{The number of reads in each bin used to calculate wcRatio}
}
\description{
Plots gapfile above histogram in plotROIFrequencies
}
\author{
Ashley D. Sanders, Mark Hills
}

