% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBam.R
\name{processBam}
\alias{processBam}
\title{Processes input bamfile}
\usage{
processBam(startLoc, endLoc, chr, fileName, padding = 1e+05, qual = 0,
  rmdup = TRUE, verbose = TRUE)
}
\arguments{
\item{startLoc}{Genomic coordinate start position}

\item{endLoc}{Genomic coordinate end position}

\item{chr}{Genomic coordinate chromosome}

\item{fileName}{Name of bamfile to process}

\item{padding}{Number of bases to extend beyond the startLoc and endLoc}

\item{qual}{Filter reads based on specified qulaity score}

\item{rmdup}{if\code{TRUE} remove duplicate reads}

\item{verbose}{Verbose messages if\code{TRUE}}
}
\description{
Generates a dataframe of reads from a bam file for the specified genomic coordinate
}
\author{
Ashley D. Sanders, Mark Hills
}

