% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBed.R
\name{processBed}
\alias{processBed}
\title{Processes input befile}
\usage{
processBed(startLoc, endLoc, chr, fileName, skip = 3, padding = 1e+05,
  qual = 0, rmdup = TRUE, verbose = TRUE)
}
\arguments{
\item{startLoc}{Genomic coordinate start position}

\item{endLoc}{Genomic coordinate end position}

\item{chr}{Genomic coordinate chromosome}

\item{fileName}{Name of bamfile to process}

\item{skip}{Number of rows in header of bedfile}

\item{padding}{Number of bases to extend beyond the startLoc and endLoc}

\item{qual}{Filter reads based on specified qulaity score}

\item{rmdup}{if\code{TRUE} remove duplicate reads}

\item{verbose}{Verbose messages if\code{TRUE}}
}
\description{
Generates a dataframe of reads from a bedfile for the specified genomic coordinate
}
\author{
Ashley D. Sanders, Mark Hills
}

