% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertRwrap.R
\name{runInvertR}
\alias{runInvertR}
\title{Wrapper function for InvertR}
\usage{
runInvertR(regionTable, binSize = 50, WCcutoff = 0.75,
  dataDirectory = "./InvertR_analysis/", gapfile = 0, type = "bed",
  dup = TRUE, qual = 10, padding = 0, minDepth = 20, minReads = 20,
  verbose = TRUE, png = TRUE, strand = TRUE, ROI = FALSE,
  genotype = TRUE, findROIs = T)
}
\arguments{
\item{regionTable}{Genomic coordinates to be analyzed (ROI list or Chr Table)}

\item{binSize}{The number of reads in each bin used to calculate wcRatio}

\item{WCcutoff}{The number of watson or crick reads used to define chrStates}

\item{dataDirectory}{Output directory. If non-existent it will be created}

\item{gapfile}{Input txt file of gaps in the genome}

\item{type}{File input type, either 'bed' or 'bam'}

\item{dup}{If \code{TRUE}, removes duplicate reads}

\item{qual}{Filter reads based on specified qulaity score}

\item{padding}{Number of bases to extend beyond genomic coordinates listed in regionTable}

\item{minDepth}{The minimum number of reads/Mb required to analyze the chromosome}

\item{minReads}{The minimum number of reads within the ROI required for inclusion}

\item{verbose}{If \code{TRUE} Verbose messages}

\item{png}{If\code{TRUE} Generates a png figure of each file}

\item{strand}{If \code{TRUE} Plot the crick and watson reads above the histrogram}

\item{ROI}{If\code{TRUE} Expects ROI list, if\code{FALSE} Expects chromosome table}

\item{findROIs}{If\code{TRUE} Runs findROIlocations to locate putative inversions}
}
\description{
This script will move through .bam or .bed files in a folder and perform several steps (see Details).
}
\details{
1. calculate the WCratio chromosome-by-chromosome
2. Locate the ROIs in chromosomes passing the WCcutoff
3. write a bedgraph file of wcRatios -> can upload on to UCSC Genome browser
4. write an ROI file for each index with all chromosomes included
}
\author{
Ashley D. Sanders, Mark Hills
}

